+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/cryptomator/hub/compare/2.0.1...HEAD)
+
+### Added
+
+- Show a short onboarding tour to newly created users, remembered across browsers and devices and replayable from the profile menu
+- Suggest downloading the Cryptomator app in the sidebar
+
 ## [2.0.1](https://github.com/cryptomator/hub/compare/2.0.0...2.0.1) (2026-09-07)
 
 ### Fixed